refactor(dialect): split functions.go along the seams it already drew
functions.go was 4381 lines and marked three of its own runs with section
comments: "PostgreSQL scalar functions", "GoogleSQL scalar functions" and
"time parsing". Files named postgresql_functions.go and googlesql_functions.go
sat beside it holding functions of exactly those dialects, so a reader looking
for PostgreSQL's TO_CHAR opened the file named for PostgreSQL and did not find
it.

The three sections move out: postgresql_helpers.go, googlesql_helpers.go and
timeparse.go. Each says in a header comment how it differs from the file next
to it -- these are reached under a rewritten name, those under the name the
query wrote -- so the split has a rule rather than a history. functions.go
keeps the registry, the coercion helpers and the common and MySQL functions,
and is 2582 lines.

No function's body changed; this is a move.
