+++ dialect/internal/runtime/functions.go
@@ -6,11 +6,9 @@
 package runtime
 
 import (
-	"crypto/md5" //nolint:gosec // MD5 backs dialects.PostgreSQL's MD5() function, not a security control
 	"crypto/rand"
 	"database/sql/driver"
 	"encoding/binary"
-	"encoding/hex"
 	"encoding/json"
 	"errors"
 	"fmt"
@@ -2580,1802 +2578,4 @@ func fnFromUnixtime(args []driver.Value) (driver.Value, error) {
 		return tm.Format(layoutDateTime), nil
 	}
 	return fnDateFormat([]driver.Value{tm.Format(layoutDateTime), args[1]})
-}
-
-// --- dialects.PostgreSQL scalar functions ---
-
-// fnToChar implements dialects.PostgreSQL TO_CHAR(value, format) for date/time values and
-// for numbers. The two are told apart by the template: a numeric one is built
-// from digit positions, which no date template contains.
-func fnToChar(args []driver.Value) (driver.Value, error) {
-	format, ok := toString(args[1])
-	if !ok {
-		return nil, nil
-	}
-	// The argument decides which template language this is, because it is the
-	// only thing that carries the answer: a digit in a date template is
-	// literal text and a letter in a numeric one is too, so dialects.PostgreSQL reads
-	// the argument's type and never the template. A number is a number; text
-	// that parses as a date is a date. Text that is neither has no type to
-	// read, and there the template is the only signal left.
-	switch value := args[0].(type) {
-	case int64:
-		return pgFormatNumber(float64(value), format), nil
-	case float64:
-		return pgFormatNumber(value, format), nil
-	}
-	if tm, ok := toStringTime(args[0]); ok {
-		return pgFormatTime(format, tm), nil
-	}
-	if !isDateTemplate(format) {
-		value, ok := toFloat(args[0])
-		if !ok {
-			return nil, nil
-		}
-		return pgFormatNumber(value, format), nil
-	}
-	return nil, nil
-}
-
-// groupThousands inserts a comma every three digits from the right.
-func groupThousands(digits string) string {
-	var b strings.Builder
-	for i, r := range digits {
-		if i > 0 && (len(digits)-i)%3 == 0 {
-			b.WriteByte(',')
-		}
-		b.WriteRune(r)
-	}
-	return b.String()
-}
-
-// fnToDate implements dialects.PostgreSQL TO_DATE(str, format).
-func fnToDate(args []driver.Value) (driver.Value, error) {
-	s, ok := toString(args[0])
-	format, ok2 := toString(args[1])
-	if !ok || !ok2 {
-		return nil, nil
-	}
-	tm, err := pgReadTemplate(format, s)
-	if err != nil {
-		return nil, err
-	}
-	return tm.Format(layoutDateOnly), nil
-}
-
-// fnDateTrunc implements dialects.PostgreSQL DATE_TRUNC(unit, timestamp).
-func fnDateTrunc(args []driver.Value) (driver.Value, error) {
-	unit, ok := toString(args[0])
-	if !ok {
-		return nil, nil
-	}
-	tm, ok := toStringTime(args[1])
-	if !ok {
-		return nil, nil
-	}
-	y, mo, d := tm.Date()
-	loc := tm.Location()
-	switch strings.ToLower(strings.TrimSpace(unit)) {
-	case unitYear:
-		return time.Date(y, 1, 1, 0, 0, 0, 0, loc).Format(layoutDateTime), nil
-	case unitQuarter:
-		q := (int(mo)-1)/3*3 + 1
-		return time.Date(y, time.Month(q), 1, 0, 0, 0, 0, loc).Format(layoutDateTime), nil
-	case unitMonth:
-		return time.Date(y, mo, 1, 0, 0, 0, 0, loc).Format(layoutDateTime), nil
-	case unitWeek:
-		offset := (int(tm.Weekday()) + 6) % 7 // days since Monday
-		monday := time.Date(y, mo, d, 0, 0, 0, 0, loc).AddDate(0, 0, -offset)
-		return monday.Format(layoutDateTime), nil
-	case unitDay:
-		return time.Date(y, mo, d, 0, 0, 0, 0, loc).Format(layoutDateTime), nil
-	case unitHour:
-		return time.Date(y, mo, d, tm.Hour(), 0, 0, 0, loc).Format(layoutDateTime), nil
-	case unitMinute:
-		return time.Date(y, mo, d, tm.Hour(), tm.Minute(), 0, 0, loc).Format(layoutDateTime), nil
-	case unitSecond:
-		return time.Date(y, mo, d, tm.Hour(), tm.Minute(), tm.Second(), 0, loc).Format(layoutDateTime), nil
-	case "decade":
-		return time.Date(decadeOf(y)*10, 1, 1, 0, 0, 0, 0, loc).Format(layoutDateTime), nil
-	case "century":
-		// A century starts in its first year, which is the year ending in 01:
-		// truncating 2024 to a century gives 2001 rather than 2000.
-		return time.Date((centuryOf(y)-1)*100+1, 1, 1, 0, 0, 0, 0, loc).Format(layoutDateTime), nil
-	case "millennium":
-		return time.Date((millenniumOf(y)-1)*1000+1, 1, 1, 0, 0, 0, 0, loc).Format(layoutDateTime), nil
-	case "milliseconds", unitMillisecond:
-		return truncatedFraction(tm, time.Millisecond), nil
-	case unitMicrosecondsPlural, unitMicrosecond:
-		return truncatedFraction(tm, time.Microsecond), nil
-	case unitISOYear:
-		// The ISO year begins on the Monday of the week holding its first
-		// Thursday, which is not January 1 in most years.
-		isoYear, _ := tm.ISOWeek()
-		jan4 := time.Date(isoYear, time.January, 4, 0, 0, 0, 0, loc)
-		offset := (int(jan4.Weekday()) + 6) % 7
-		return jan4.AddDate(0, 0, -offset).Format(layoutDateTime), nil
-	default:
-		return nil, fmt.Errorf("dialect: unsupported DATE_TRUNC unit %q", unit)
-	}
-}
-
-// truncatedFraction rounds a time down to a multiple of unit and spells it with
-// however much of the fraction survives, trailing zeros removed, which is how
-// dialects.PostgreSQL prints a timestamp: DATE_TRUNC('millisecond', '10:11:12.123456')
-// is 10:11:12.123 and a whole second keeps no decimal point at all.
-func truncatedFraction(tm time.Time, unit time.Duration) string {
-	out := tm.Truncate(unit)
-	if out.Nanosecond() == 0 {
-		return out.Format(layoutDateTime)
-	}
-	return strings.TrimRight(out.Format(layoutDateTime+".000000000"), "0")
-}
-
-// fnSplitPart implements dialects.PostgreSQL SPLIT_PART(string, delimiter, n) with a
-// 1-based field index.
-func fnSplitPart(args []driver.Value) (driver.Value, error) {
-	s, ok1 := toString(args[0])
-	delim, ok2 := toString(args[1])
-	n, ok3 := toInt(args[2])
-	if !ok1 || !ok2 || !ok3 {
-		return nil, nil
-	}
-	// dialects.PostgreSQL refuses a zero field position rather than answering with an
-	// empty string, which is what makes an off-by-one in a computed position
-	// visible instead of reading as an empty field.
-	if n == 0 {
-		return nil, errors.New("dialect: SPLIT_PART: field position must not be zero")
-	}
-	if delim == "" {
-		if n == 1 || n == -1 {
-			return s, nil
-		}
-		return "", nil
-	}
-	parts := strings.Split(s, delim)
-	idx := int(n)
-	if idx < 0 {
-		idx = len(parts) + idx + 1
-	}
-	if idx < 1 || idx > len(parts) {
-		return "", nil
-	}
-	return parts[idx-1], nil
-}
-
-// fnInitcap implements dialects.PostgreSQL INITCAP: uppercase the first letter of each
-// alphanumeric run, lowercase the rest.
-func fnInitcap(args []driver.Value) (driver.Value, error) {
-	s, ok := toString(args[0])
-	if !ok {
-		return nil, nil
-	}
-	var b strings.Builder
-	prevAlnum := false
-	for _, r := range s {
-		alnum := isAlnumRune(r)
-		switch {
-		case alnum && !prevAlnum:
-			b.WriteString(strings.ToUpper(string(r)))
-		case alnum:
-			b.WriteString(strings.ToLower(string(r)))
-		default:
-			b.WriteRune(r)
-		}
-		prevAlnum = alnum
-	}
-	return b.String(), nil
-}
-
-// isAlnumRune reports whether r is part of a word, which is what decides where
-// INITCAP capitalizes. A letter is a letter whatever its script: testing the
-// ASCII range read an accented letter as a separator, so "école" came back
-// "éCole".
-// fnUnicodeUpper and fnUnicodeLower fold case over the whole of Unicode, which
-// is what dialects.MySQL, dialects.PostgreSQL and dialects.GoogleSQL do and what dialects.SQLite's own upper() and
-// lower() do not: theirs stop at ASCII, so UPPER('école') came back 'éCOLE'.
-func fnUnicodeUpper(args []driver.Value) (driver.Value, error) {
-	s, ok := toString(args[0])
-	if !ok {
-		return nil, nil
-	}
-	return strings.ToUpper(s), nil
-}
-
-func fnUnicodeLower(args []driver.Value) (driver.Value, error) {
-	s, ok := toString(args[0])
-	if !ok {
-		return nil, nil
-	}
-	return strings.ToLower(s), nil
-}
-
-func isAlnumRune(r rune) bool {
-	return unicode.IsLetter(r) || unicode.IsDigit(r)
-}
-
-// fnStrpos implements dialects.PostgreSQL STRPOS(string, substring): 1-based index or 0.
-func fnStrpos(args []driver.Value) (driver.Value, error) {
-	s, ok1 := toString(args[0])
-	sub, ok2 := toString(args[1])
-	if !ok1 || !ok2 {
-		return nil, nil
-	}
-	return int64(characterIndex(s, sub, 0)), nil
-}
-
-// characterIndex is the 1-based position of sub in s counted in characters, or
-// 0 when it is not there. from is a character offset to start at, which is what
-// LOCATE's third argument means. Every dialect here counts a position in
-// characters; strings.Index answers in bytes, and returning that answered a
-// number that indexes nothing in text outside ASCII.
-func characterIndex(s, sub string, from int) int {
-	runes := []rune(s)
-	if from < 0 || from > len(runes) {
-		return 0
-	}
-	tail := string(runes[from:])
-	idx := strings.Index(tail, sub)
-	if idx < 0 {
-		return 0
-	}
-	return from + utf8.RuneCountInString(tail[:idx]) + 1
-}
-
-func fnLeft(args []driver.Value) (driver.Value, error)  { return leftRight(args, true) }
-func fnRight(args []driver.Value) (driver.Value, error) { return leftRight(args, false) }
-
-func fnMySQLLeft(args []driver.Value) (driver.Value, error)  { return mysqlLeftRight(args, true) }
-func fnMySQLRight(args []driver.Value) (driver.Value, error) { return mysqlLeftRight(args, false) }
-
-func fnGoogleSQLLeft(args []driver.Value) (driver.Value, error) {
-	return googlesqlLeftRight(args, true)
-}
-
-func fnGoogleSQLRight(args []driver.Value) (driver.Value, error) {
-	return googlesqlLeftRight(args, false)
-}
-
-// leftRight implements LEFT/RIGHT with dialects.PostgreSQL's negative-count semantics: a
-// negative n removes |n| characters from the far end.
-func leftRight(args []driver.Value, left bool) (driver.Value, error) {
-	s, n, ok := leftRightArgs(args)
-	if !ok {
-		return nil, nil
-	}
-	count := int(n)
-	if count < 0 {
-		count = len([]rune(s)) + count
-	}
-	return takeRunes(s, count, left), nil
-}
-
-// mysqlLeftRight implements LEFT/RIGHT with dialects.MySQL's negative-count semantics: a
-// negative n answers the empty string rather than trimming the far end.
-func mysqlLeftRight(args []driver.Value, left bool) (driver.Value, error) {
-	s, n, ok := leftRightArgs(args)
-	if !ok {
-		return nil, nil
-	}
-	return takeRunes(s, int(n), left), nil
-}
-
-// googlesqlLeftRight implements LEFT/RIGHT with dialects.GoogleSQL's rule for a negative
-// count, which is to raise: BigQuery has no meaning for a negative length and
-// answering dialects.PostgreSQL's trimmed string for one would hide the mistake.
-func googlesqlLeftRight(args []driver.Value, left bool) (driver.Value, error) {
-	s, n, ok := leftRightArgs(args)
-	if !ok {
-		return nil, nil
-	}
-	if n < 0 {
-		return nil, fmt.Errorf("dialect: LEFT/RIGHT length must not be negative, got %d", n)
-	}
-	return takeRunes(s, int(n), left), nil
-}
-
-// leftRightArgs coerces the shared (string, count) arguments of LEFT and RIGHT.
-func leftRightArgs(args []driver.Value) (string, int64, bool) {
-	s, ok1 := toString(args[0])
-	n, ok2 := toCount(args[1])
-	return s, n, ok1 && ok2
-}
-
-// takeRunes returns the first or last count characters of s, the whole of s when
-// it is shorter than that, and the empty string when count is not positive.
-func takeRunes(s string, count int, fromLeft bool) string {
-	if count <= 0 {
-		return ""
-	}
-	runes := []rune(s)
-	if count > len(runes) {
-		count = len(runes)
-	}
-	if fromLeft {
-		return string(runes[:count])
-	}
-	return string(runes[len(runes)-count:])
-}
-
-// fnRegexpReplace implements dialects.GoogleSQL REGEXP_REPLACE(source, pattern,
-// replacement), which replaces every match, and is also what a query written in
-// the dialects.SQLite dialect reaches. dialects.PostgreSQL back-references (\1) are translated to
-// Go's ${1} expansion form. A flags argument is accepted here for the callers
-// that already pass one: "g" replaces every match and its absence replaces only
-// the first, and "i" matches case insensitively.
-func fnRegexpReplace(args []driver.Value) (driver.Value, error) {
-	return regexpReplace(args, true)
-}
-
-// fnPostgresRegexpReplace implements dialects.PostgreSQL regexp_replace(source, pattern,
-// replacement[, flags]), whose three-argument form replaces the first match
-// alone; the rest need the "g" flag.
-func fnPostgresRegexpReplace(args []driver.Value) (driver.Value, error) {
-	return regexpReplace(args, false)
-}
-
-// regexpReplace is the shared body of the flag-taking REGEXP_REPLACE forms.
-// defaultGlobal is what the three-argument form means, which is every match for
-// dialects.GoogleSQL and the first alone for dialects.PostgreSQL.
-func regexpReplace(args []driver.Value, defaultGlobal bool) (driver.Value, error) {
-	if len(args) < 3 || len(args) > 4 {
-		return nil, fmt.Errorf("dialect: REGEXP_REPLACE expects 3 or 4 arguments, got %d", len(args))
-	}
-	src, ok1 := toString(args[0])
-	pattern, ok2 := toString(args[1])
-	repl, ok3 := toString(args[2])
-	if !ok1 || !ok2 || !ok3 {
-		return nil, nil
-	}
-	global := defaultGlobal
-	if len(args) == 4 {
-		flags, ok := toString(args[3])
-		if !ok {
-			return nil, nil
-		}
-		global = strings.Contains(flags, "g")
-		if strings.Contains(flags, "i") {
-			pattern = "(?i)" + pattern
-		}
-	}
-	re, err := compileRegexp(pattern)
-	if err != nil {
-		return nil, err
-	}
-	expansion := pgReplacement(repl)
-	if global {
-		return re.ReplaceAllString(src, expansion), nil
-	}
-	// Expand against the source rather than the matched text on its own: a
-	// boundary-dependent pattern such as `\Bb` matches inside "ab" but not in
-	// the isolated "b", which would leave ExpandString without submatch indices.
-	loc := re.FindStringSubmatchIndex(src)
-	if loc == nil {
-		return src, nil
-	}
-	out := re.ExpandString([]byte(src[:loc[0]]), expansion, src, loc)
-	return string(out) + src[loc[1]:], nil
-}
-
-// fnMySQLRegexpReplace implements dialects.MySQL REGEXP_REPLACE(subject, pattern,
-// replacement[, pos[, occurrence[, match_type]]]). dialects.MySQL's fourth argument is a
-// 1-based character position to start at rather than dialects.PostgreSQL's flag string,
-// and its fifth selects one match to replace, with 0 meaning every match from
-// the position onward.
-func fnMySQLRegexpReplace(args []driver.Value) (driver.Value, error) {
-	if len(args) < 3 || len(args) > 6 {
-		return nil, fmt.Errorf("dialect: REGEXP_REPLACE expects 3 to 6 arguments, got %d", len(args))
-	}
-	src, ok1 := toString(args[0])
-	pattern, ok2 := toString(args[1])
-	repl, ok3 := toString(args[2])
-	if !ok1 || !ok2 || !ok3 {
-		return nil, nil
-	}
-	pos, occurrence := int64(1), int64(0)
-	if len(args) >= 4 {
-		n, ok := toInt(args[3])
-		if !ok {
-			return nil, nil
-		}
-		pos = n
-	}
-	if len(args) >= 5 {
-		n, ok := toInt(args[4])
-		if !ok {
-			return nil, nil
-		}
-		occurrence = n
-	}
-	matchType := ""
-	if len(args) == 6 {
-		m, ok := toString(args[5])
-		if !ok {
-			return nil, nil
-		}
-		matchType = m
-	}
-	// The pattern goes through the match-type mapping even when no match type
-	// was given, because dialects.MySQL's default is not Go's: its collation folds case.
-	pattern, err := mysqlRegexpPattern(pattern, matchType)
-	if err != nil {
-		return nil, err
-	}
-	runes := []rune(src)
-	if pos < 1 || int(pos) > len(runes)+1 {
-		return nil, fmt.Errorf("dialect: REGEXP_REPLACE position %d is out of bounds", pos)
-	}
-	re, err := compileRegexp(pattern)
-	if err != nil {
-		return nil, err
-	}
-	head, tail := string(runes[:pos-1]), string(runes[pos-1:])
-	expansion := mysqlReplacement(repl)
-	if occurrence == 0 {
-		return head + re.ReplaceAllString(tail, expansion), nil
-	}
-	// A negative occurrence is not a form dialects.MySQL documents; it answers the first
-	// match there, so the count below starts at one for anything under it.
-	wanted := int(occurrence)
-	if wanted < 1 {
-		wanted = 1
-	}
-	matches := re.FindAllStringSubmatchIndex(tail, wanted)
-	if len(matches) < wanted {
-		return src, nil
-	}
-	loc := matches[wanted-1]
-	out := re.ExpandString([]byte(tail[:loc[0]]), expansion, tail, loc)
-	return head + string(out) + tail[loc[1]:], nil
-}
-
-// applyMySQLMatchType folds a dialects.MySQL match_type string into the pattern as Go
-// regexp flags. dialects.MySQL spells them c (case sensitive), i (case insensitive), m
-// (multi-line) and n (a dot matches a newline); u, which selects Unix line
-// endings, has no Go equivalent and is refused rather than ignored.
-func mysqlRegexpPattern(pattern, matchType string) (string, error) {
-	// MySQL refuses an empty pattern rather than matching everywhere with it,
-	// which is what Go's regexp does and what this answered: REGEXP_LIKE(x, '')
-	// was true for every row, including the rows a caller wrote the query to
-	// find.
-	if pattern == "" {
-		return "", errors.New("dialect: the regular expression is empty, which MySQL refuses")
-	}
-	fold := true
-	var flags string
-	for _, c := range matchType {
-		switch c {
-		case 'c':
-			fold = false
-		case 'i':
-			fold = true
-		case 'm':
-			flags += "m"
-		case 'n':
-			flags += "s"
-		default:
-			return "", fmt.Errorf("dialect: regular expression match type %q is not supported", matchType)
-		}
-	}
-	if fold {
-		flags = "i" + flags
-	} else {
-		flags += "-i"
-	}
-	// The flag group leads the pattern, so a group the caller wrote themselves
-	// stands after it and wins for the rest of the pattern.
-	return "(?" + flags + ")" + pattern, nil
-}
-
-// mysqlReplacement translates dialects.MySQL replacement references ($1..$9) to the ${n}
-// form Go's regexp expansion understands. dialects.MySQL writes a literal "$" as "\$",
-// and a backslash before anything else stands for that character.
-func mysqlReplacement(repl string) string {
-	var b strings.Builder
-	for i := 0; i < len(repl); i++ {
-		switch {
-		case repl[i] == '\\' && i+1 < len(repl):
-			if repl[i+1] == '$' {
-				b.WriteString("$$")
-			} else {
-				b.WriteByte(repl[i+1])
-			}
-			i++
-		case repl[i] == '$' && i+1 < len(repl) && repl[i+1] >= '0' && repl[i+1] <= '9':
-			b.WriteString("${")
-			b.WriteByte(repl[i+1])
-			b.WriteByte('}')
-			i++
-		case repl[i] == '$':
-			b.WriteString("$$")
-		default:
-			b.WriteByte(repl[i])
-		}
-	}
-	return b.String()
-}
-
-// fnMD5 implements dialects.PostgreSQL MD5(text). MD5 is used here as a content
-// fingerprint compatible with the source dialect, never for security.
-func fnMD5(args []driver.Value) (driver.Value, error) {
-	s, ok := toString(args[0])
-	if !ok {
-		return nil, nil
-	}
-	sum := md5.Sum([]byte(s)) //nolint:gosec // required for dialects.PostgreSQL MD5() compatibility, not a security control
-	return hex.EncodeToString(sum[:]), nil
-}
-
-// fnASCII implements ASCII(text): the code point of the first character, or 0
-// for an empty string.
-func fnASCII(args []driver.Value) (driver.Value, error) {
-	s, ok := toString(args[0])
-	if !ok {
-		return nil, nil
-	}
-	for _, r := range s {
-		return int64(r), nil
-	}
-	return int64(0), nil
-}
-
-// fnChr implements CHR(code): the character for a code point.
-func fnChr(args []driver.Value) (driver.Value, error) {
-	n, ok := toInt(args[0])
-	if !ok {
-		return nil, nil
-	}
-	if n <= 0 || n > utf8.MaxRune {
-		// dialects.PostgreSQL refuses zero as well as a negative code point: a zero byte
-		// cannot be stored in a text value there. Answering the space that
-		// dialects.SQLite's char() leaves behind for it is the one result a caller
-		// cannot tell from a real character.
-		return nil, fmt.Errorf("dialect: CHR: code point %d is out of range", n)
-	}
-	return string(rune(n)), nil
-}
-
-// fnTranslate implements dialects.PostgreSQL TRANSLATE(string, from, to): each character
-// of from is replaced by the character at the same position in to, and a
-// character whose position has no counterpart in to is dropped.
-func fnTranslate(args []driver.Value) (driver.Value, error) {
-	s, ok1 := toString(args[0])
-	from, ok2 := toString(args[1])
-	to, ok3 := toString(args[2])
-	if !ok1 || !ok2 || !ok3 {
-		return nil, nil
-	}
-	fromRunes := []rune(from)
-	toRunes := []rune(to)
-	var b strings.Builder
-	for _, r := range s {
-		idx := -1
-		for i, f := range fromRunes {
-			if f == r {
-				idx = i
-				break
-			}
-		}
-		switch {
-		case idx < 0:
-			b.WriteRune(r)
-		case idx < len(toRunes):
-			b.WriteRune(toRunes[idx])
-		}
-	}
-	return b.String(), nil
-}
-
-// pgReplacement translates dialects.PostgreSQL replacement back-references (\1..\9, \&) to
-// the ${n} form Go's regexp expansion understands.
-func pgReplacement(repl string) string {
-	var b strings.Builder
-	for i := 0; i < len(repl); i++ {
-		if repl[i] == '\\' && i+1 < len(repl) {
-			c := repl[i+1]
-			switch {
-			case c >= '0' && c <= '9':
-				b.WriteString("${")
-				b.WriteByte(c)
-				b.WriteByte('}')
-			case c == '&':
-				b.WriteString("${0}")
-			default:
-				b.WriteByte(c)
-			}
-			i++
-			continue
-		}
-		if repl[i] == '$' {
-			// Escape a literal '$' so Go does not treat it as an expansion.
-			b.WriteString("$$")
-			continue
-		}
-		b.WriteByte(repl[i])
-	}
-	return b.String()
-}
-
-// --- dialects.GoogleSQL scalar functions ---
-
-// fnSafeDivide implements dialects.GoogleSQL SAFE_DIVIDE(x, y): x/y, or NULL when y is 0
-// or either argument is NULL.
-func fnSafeDivide(args []driver.Value) (driver.Value, error) {
-	x, ok1 := toFloat(args[0])
-	y, ok2 := toFloat(args[1])
-	if !ok1 || !ok2 || y == 0 {
-		return nil, nil
-	}
-	return x / y, nil
-}
-
-// fnStartsWith implements dialects.GoogleSQL STARTS_WITH(value, prefix).
-func fnStartsWith(args []driver.Value) (driver.Value, error) {
-	s, ok1 := toString(args[0])
-	prefix, ok2 := toString(args[1])
-	if !ok1 || !ok2 {
-		return nil, nil
-	}
-	return boolToInt(strings.HasPrefix(s, prefix)), nil
-}
-
-// fnEndsWith implements dialects.GoogleSQL ENDS_WITH(value, suffix).
-func fnEndsWith(args []driver.Value) (driver.Value, error) {
-	s, ok1 := toString(args[0])
-	suffix, ok2 := toString(args[1])
-	if !ok1 || !ok2 {
-		return nil, nil
-	}
-	return boolToInt(strings.HasSuffix(s, suffix)), nil
-}
-
-func boolToInt(b bool) int64 {
-	if b {
-		return 1
-	}
-	return 0
-}
-
-// fnRegexpContains implements dialects.GoogleSQL REGEXP_CONTAINS(value, pattern). Note the
-// argument order is (value, pattern), the reverse of the REGEXP operator.
-func fnRegexpContains(args []driver.Value) (driver.Value, error) {
-	value, ok1 := toString(args[0])
-	pattern, ok2 := toString(args[1])
-	if !ok1 || !ok2 {
-		return nil, nil
-	}
-	re, err := compileRegexp(pattern)
-	if err != nil {
-		return nil, err
-	}
-	return boolToInt(re.MatchString(value)), nil
-}
-
-// fnRegexpExtract implements dialects.GoogleSQL REGEXP_EXTRACT(value, pattern): the first
-// capturing group when the pattern has one, otherwise the whole match, or NULL
-// when there is no match.
-func fnRegexpExtract(args []driver.Value) (driver.Value, error) {
-	value, ok1 := toString(args[0])
-	pattern, ok2 := toString(args[1])
-	if !ok1 || !ok2 {
-		return nil, nil
-	}
-	re, err := compileRegexp(pattern)
-	if err != nil {
-		return nil, err
-	}
-	m := re.FindStringSubmatch(value)
-	if m == nil {
-		return nil, nil
-	}
-	if len(m) > 1 {
-		return m[1], nil
-	}
-	return m[0], nil
-}
-
-// fnDateDiff3 implements dialects.GoogleSQL DATE_DIFF/TIMESTAMP_DIFF(a, b, unit): the
-// signed count of unit boundaries from b to a.
-func fnDateDiff3(args []driver.Value) (driver.Value, error) {
-	a, ok1 := toStringTime(args[0])
-	b, ok2 := toStringTime(args[1])
-	unit, ok3 := toString(args[2])
-	if !ok1 || !ok2 || !ok3 {
-		return nil, nil
-	}
-	switch part := strings.ToLower(strings.TrimSpace(unit)); part {
-	case unitYear:
-		return int64(a.Year() - b.Year()), nil
-	case unitISOYear:
-		yearA, _ := a.ISOWeek()
-		yearB, _ := b.ISOWeek()
-		return int64(yearA - yearB), nil
-	case unitQuarter:
-		return int64((a.Year()*4 + (int(a.Month())-1)/3) - (b.Year()*4 + (int(b.Month())-1)/3)), nil
-	case unitMonth:
-		return int64((a.Year()*12 + int(a.Month())) - (b.Year()*12 + int(b.Month()))), nil
-	case unitWeek, unitISOWeek:
-		// A week difference counts the week boundaries crossed rather than the
-		// seven-day spans between the two dates, so a Saturday and the Sunday
-		// after it are one week apart. WEEK begins on Sunday and ISOWEEK on
-		// Monday.
-		start := time.Sunday
-		if part == unitISOWeek {
-			start = time.Monday
-		}
-		return int64(weekBoundariesBetween(a, b, start)), nil
-	case unitDay:
-		return dayNumber(a) - dayNumber(b), nil
-	case unitHour:
-		return secondsBetween(a, b) / 3600, nil
-	case unitMinute:
-		return secondsBetween(a, b) / 60, nil
-	case unitSecond:
-		return secondsBetween(a, b), nil
-	case unitMillisecond:
-		return a.Sub(b).Milliseconds(), nil
-	case unitMicrosecond:
-		return a.Sub(b).Microseconds(), nil
-	default:
-		if start, ok := weekStartDay(part); ok {
-			return int64(weekBoundariesBetween(a, b, start)), nil
-		}
-		return nil, fmt.Errorf("dialect: unsupported DATE_DIFF unit %q", unit)
-	}
-}
-
-// weekBoundariesBetween counts the weeks that begin on start between two dates,
-// signed, which is what DATE_DIFF counts for every week spelling.
-func weekBoundariesBetween(a, b time.Time, start time.Weekday) int {
-	weekStartDayNumber := func(t time.Time) int64 {
-		return dayNumber(t) - int64((int(t.Weekday())-int(start)+7)%7)
-	}
-	return int((weekStartDayNumber(a) - weekStartDayNumber(b)) / 7)
-}
-
-// dayNumber is the day a time falls on, counted from 1970-01-01 and computed
-// from the civil date rather than from a duration. A time.Duration is an int64
-// of nanoseconds, so subtracting two times saturates about 292 years out and
-// answers the bound rather than the distance -- which made the days to a
-// 9999-12-31 sentinel 106751 instead of 2913109.
-func dayNumber(t time.Time) int64 {
-	y, m, d := t.Date()
-	return civilDayNumber(int64(y), int64(m), int64(d))
-}
-
-// civilDayNumber is the day number of a proleptic Gregorian date, counted from
-// 1970-01-01. It shifts the year to start in March so a leap day falls at the
-// end of a four-century cycle, which is what lets the whole thing be integer
-// arithmetic with no table and no bound short of the int64 the year is in.
-func civilDayNumber(year, month, day int64) int64 {
-	if month <= 2 {
-		year--
-	}
-	era := year / 400
-	if year < 0 {
-		era = (year - 399) / 400
-	}
-	yearOfEra := year - era*400 // [0, 399]
-	monthShift := int64(9)
-	if month > 2 {
-		monthShift = -3
-	}
-	dayOfYear := (153*(month+monthShift)+2)/5 + day - 1                 // [0, 365]
-	dayOfEra := yearOfEra*365 + yearOfEra/4 - yearOfEra/100 + dayOfYear // [0, 146096]
-	return era*146097 + dayOfEra - 719468
-}
-
-// secondsBetween is the signed distance between two times in seconds, built
-// from the day number and the time of day so it is exact for every date the
-// calendar holds.
-func secondsBetween(a, b time.Time) int64 {
-	const secondsPerDay = 24 * 60 * 60
-	days := dayNumber(a) - dayNumber(b)
-	return days*secondsPerDay + secondsOfDay(a) - secondsOfDay(b)
-}
-
-// secondsOfDay is how far into its day a time is, in whole seconds.
-func secondsOfDay(t time.Time) int64 {
-	return int64(t.Hour())*3600 + int64(t.Minute())*60 + int64(t.Second())
-}
-
-// fnMySQLDateDiff implements dialects.MySQL TIMESTAMPDIFF(unit, start, end), called as
-// mysql_date_diff(end, start, 'unit'): the signed number of complete units from
-// start to end, where fnDateDiff3 counts the boundaries BigQuery counts. DAY
-// and WEEK are complete 24-hour and 7-day periods of the actual span, so 23
-// hours is zero days; MONTH, QUARTER and YEAR are complete calendar months.
-func fnMySQLDateDiff(args []driver.Value) (driver.Value, error) {
-	end, ok1 := toStringTime(args[0])
-	start, ok2 := toStringTime(args[1])
-	unit, ok3 := toString(args[2])
-	if !ok1 || !ok2 || !ok3 {
-		return nil, nil
-	}
-	switch strings.ToLower(strings.TrimSpace(unit)) {
-	case unitMicrosecond:
-		return end.Sub(start).Microseconds(), nil
-	case unitSecond:
-		return wholeUnits(end, start, 1), nil
-	case unitMinute:
-		return wholeUnits(end, start, 60), nil
-	case unitHour:
-		return wholeUnits(end, start, 3600), nil
-	case unitDay:
-		return wholeUnits(end, start, 24*3600), nil
-	case unitWeek:
-		return wholeUnits(end, start, daysPerWeek*24*3600), nil
-	case unitMonth:
-		return completeMonths(end, start), nil
-	case unitQuarter:
-		return completeMonths(end, start) / monthsPerQuartr, nil
-	case unitYear:
-		return completeMonths(end, start) / 12, nil
-	default:
-		return nil, fmt.Errorf("dialect: unsupported TIMESTAMPDIFF unit %q", unit)
-	}
-}
-
-// wholeUnits is the signed number of complete unitSeconds-second units from
-// start to end, truncated toward zero. It counts in whole seconds rather than
-// through a time.Duration, which saturates at about ±292 years while dialects.MySQL's
-// DATETIME range spans nine millennia.
-func wholeUnits(end, start time.Time, unitSeconds int64) int64 {
-	seconds := end.Unix() - start.Unix()
-	// The nanosecond fields decide whether the span's last second is complete.
-	if seconds > 0 && end.Nanosecond() < start.Nanosecond() {
-		seconds--
-	} else if seconds < 0 && end.Nanosecond() > start.Nanosecond() {
-		seconds++
-	}
-	return seconds / unitSeconds
-}
-
-// completeMonths is the signed number of complete calendar months from start to
-// end: the month delta, minus one when the end's day and time of day fall short
-// of the start's. dialects.MySQL has no month-end special case, so January 31 to
-// February 29 is zero complete months even though February 29 ends its month.
-func completeMonths(end, start time.Time) int64 {
-	sign := int64(1)
-	if end.Before(start) {
-		end, start = start, end
-		sign = -1
-	}
-	months := int64((end.Year()-start.Year())*12 + int(end.Month()) - int(start.Month()))
-	if end.Day() < start.Day() || (end.Day() == start.Day() && clockNanos(end) < clockNanos(start)) {
-		months--
-	}
-	return sign * months
-}
-
-// clockNanos is the time of day as nanoseconds since midnight.
-func clockNanos(tm time.Time) int64 {
-	return int64(tm.Hour())*int64(time.Hour) + int64(tm.Minute())*int64(time.Minute) +
-		int64(tm.Second())*int64(time.Second) + int64(tm.Nanosecond())
-}
-
-// mysqlTimeMaxSeconds is the ceiling of dialects.MySQL's TIME type, 838:59:59, which
-// TIMEDIFF clamps to on both sides.
-const mysqlTimeMaxSeconds = 838*3600 + 59*60 + 59
-
-// fnMySQLTimeDiff implements dialects.MySQL TIMEDIFF(a, b): a minus b rendered as a
-// dialects.MySQL TIME, whose hours keep counting past 24 and clamp at ±838:59:59.
-// Two datetimes and two bare TIME values both work; mixing the two shapes is
-// NULL, as it is in dialects.MySQL, and so is a value that does not parse.
-func fnMySQLTimeDiff(args []driver.Value) (driver.Value, error) {
-	s1, ok1 := toString(args[0])
-	s2, ok2 := toString(args[1])
-	if !ok1 || !ok2 {
-		return nil, nil
-	}
-	if hasCalendarDate(s1) != hasCalendarDate(s2) {
-		return nil, nil
-	}
-	if !hasCalendarDate(s1) {
-		// Bare TIME values, whose hours may pass 23 and carry a sign, which
-		// the calendar parser cannot read.
-		n1, ok1 := mysqlClockNanos(s1)
-		n2, ok2 := mysqlClockNanos(s2)
-		if !ok1 || !ok2 {
-			return nil, nil
-		}
-		return renderMySQLTime(n1 - n2), nil
-	}
-	t1, ok1 := parseTime(s1)
-	t2, ok2 := parseTime(s2)
-	if !ok1 || !ok2 {
-		return nil, nil
-	}
-	// Sub saturates at ±292 years, far past the ±838:59:59 clamp, so the
-	// saturated value renders the same clamped TIME.
-	return renderMySQLTime(int64(t1.Sub(t2))), nil
-}
-
-// renderMySQLTime renders signed nanoseconds as a dialects.MySQL TIME, clamped to
-// ±838:59:59, with a six-digit fraction only when the value has one.
-func renderMySQLTime(nanos int64) string {
-	sign := ""
-	if nanos < 0 {
-		sign = "-"
-		nanos = -nanos
-	}
-	if nanos > mysqlTimeMaxSeconds*int64(time.Second) {
-		nanos = mysqlTimeMaxSeconds * int64(time.Second)
-	}
-	secs := nanos / int64(time.Second)
-	out := fmt.Sprintf("%s%02d:%02d:%02d", sign, secs/3600, secs/60%60, secs%60)
-	if frac := nanos % int64(time.Second); frac != 0 {
-		out += fmt.Sprintf(".%06d", frac/int64(time.Microsecond))
-	}
-	return out
-}
-
-// mysqlClockNanos reads a dialects.MySQL TIME value — an optional sign, hours that may
-// pass 23, minutes, seconds and an optional fraction — as signed nanoseconds
-// from 00:00:00. ok is false when s is not that shape.
-func mysqlClockNanos(s string) (int64, bool) {
-	s = strings.TrimSpace(s)
-	rest, negative := strings.CutPrefix(s, "-")
-	clock, frac, hasFrac := strings.Cut(rest, ".")
-	parts := strings.Split(clock, ":")
-	if len(parts) != 3 {
-		return 0, false
-	}
-	fields := make([]int64, 3)
-	for i, part := range parts {
-		if part == "" || (i > 0 && len(part) > 2) {
-			return 0, false
-		}
-		n, err := strconv.ParseInt(part, 10, 64)
-		if err != nil {
-			return 0, false
-		}
-		fields[i] = n
-	}
-	if fields[1] > 59 || fields[2] > 59 {
-		return 0, false
-	}
-	// dialects.MySQL clamps each TIME argument to its range before subtracting, so
-	// TIMEDIFF('2000:00:00', '1000:00:00') is zero. Clamping here also keeps
-	// an absurd hour count from overflowing the multiplication.
-	if fields[0] > mysqlTimeMaxSeconds/3600 {
-		nanos := int64(mysqlTimeMaxSeconds) * int64(time.Second)
-		if negative {
-			nanos = -nanos
-		}
-		return nanos, true
-	}
-	nanos := (fields[0]*3600 + fields[1]*60 + fields[2]) * int64(time.Second)
-	if hasFrac {
-		if len(frac) > 9 {
-			frac = frac[:9]
-		}
-		n, err := strconv.ParseInt(frac+strings.Repeat("0", 9-len(frac)), 10, 64)
-		if err != nil {
-			return 0, false
-		}
-		nanos += n
-	}
-	if negative {
-		nanos = -nanos
-	}
-	return nanos, true
-}
-
-// hasCalendarDate reports whether a datetime string carries a calendar date,
-// which is what separates '2024-01-01 01:00:00' from a bare '01:00:00'. A
-// leading minus is a TIME's sign, not a date separator.
-func hasCalendarDate(s string) bool {
-	return strings.ContainsAny(strings.TrimPrefix(strings.TrimSpace(s), "-"), "-/")
-}
-
-// strftimeToGoLayout maps the strftime-style specifiers dialects.GoogleSQL uses in
-// FORMAT_DATE and PARSE_DATE to Go reference-time layout fragments. They differ
-// from the dialects.MySQL DATE_FORMAT set: %M is the minute here but the month name
-// there, and the month and weekday names are %B and %A.
-var strftimeToGoLayout = map[byte]string{
-	'Y': "2006",
-	'y': "06",
-	'm': "01",
-	'd': "02",
-	'e': "_2",
-	'H': "15",
-	'I': "03",
-	'M': "04",
-	'S': "05",
-	'p': "PM",
-	'B': layoutMonthLong,
-	'b': layoutMonthShort,
-	'A': layoutWeekdayLong,
-	'a': layoutWeekdayShort,
-	'F': layoutDateOnly,
-	'T': layoutTimeOnly,
-	'R': "15:04",
-	'h': layoutMonthShort,
-	'P': "pm",
-	'D': "01/02/06",
-	'x': "01/02/06",
-	'X': layoutTimeOnly,
-	'c': "Mon Jan _2 15:04:05 2006",
-}
-
-// strftimeLayout converts a dialects.GoogleSQL format string into a Go layout. An
-// unknown specifier contributes its own letter, mirroring how DATE_FORMAT
-// handles the same case.
-func strftimeLayout(format string) string {
-	var b strings.Builder
-	for i := 0; i < len(format); i++ {
-		if format[i] == '%' && i+1 < len(format) {
-			if layout, ok := strftimeToGoLayout[format[i+1]]; ok {
-				b.WriteString(layout)
-			} else if format[i+1] == '%' {
-				b.WriteByte('%')
-			} else {
-				b.WriteByte(format[i+1])
-			}
-			i++
-			continue
-		}
-		b.WriteByte(format[i])
-	}
-	return b.String()
-}
-
-// strftimeRender writes tm according to a dialects.GoogleSQL format string.
-//
-// Rendering is separate from strftimeLayout, which builds a Go layout for
-// PARSE_DATE, because a layout can only carry what Go has a reference-time
-// fragment for. A day of the year, a week number and an epoch second are
-// computed rather than spelled, so twenty specifiers had nowhere to go and were
-// written as their own letter -- which is what BigQuery does for a specifier it
-// does not know, so a format asking for a time came back holding an "X" and
-// looked like it had worked.
-//
-// Building the answer rather than a layout also stops the text around the
-// specifiers being read as one. A literal "2006" or "1" in a format string used
-// to reach time.Format, which rendered it as the year or the month; the
-// characters between specifiers are copied verbatim now.
-func strftimeRender(tm time.Time, format string) string {
-	var b strings.Builder
-	for i := 0; i < len(format); i++ {
-		if format[i] != '%' || i+1 >= len(format) {
-			b.WriteByte(format[i])
-			continue
-		}
-		spec := format[i+1]
-		i++
-		if spec == '%' {
-			b.WriteByte('%')
-			continue
-		}
-		if layout, ok := strftimeToGoLayout[spec]; ok {
-			b.WriteString(tm.Format(layout))
-			continue
-		}
-		if computed, ok := strftimeComputed(tm, spec); ok {
-			b.WriteString(computed)
-			continue
-		}
-		// An unknown specifier is its own letter, which is what BigQuery does.
-		b.WriteByte(spec)
-	}
-	return b.String()
-}
-
-// strftimeComputed handles the specifiers with no Go layout fragment: the ones
-// whose value is calculated from the date rather than spelled out of it.
-//
-// The week numbers are the four strftime defines. %V and %G are the ISO pair Go
-// answers directly; %U counts weeks from the year's first Sunday and %W from its
-// first Monday, both numbering the days before that week 0, which is the
-// calculation the C library does and is written out here rather than derived
-// from the ISO pair, since the three disagree at every turn of the year.
-func strftimeComputed(tm time.Time, spec byte) (string, bool) {
-	switch spec {
-	case 'j':
-		return fmt.Sprintf("%03d", tm.YearDay()), true
-	case 's':
-		return strconv.FormatInt(tm.Unix(), 10), true
-	case 'C':
-		return fmt.Sprintf("%02d", tm.Year()/100), true
-	case 'Q':
-		return strconv.Itoa((int(tm.Month())-1)/3 + 1), true
-	case 'k':
-		return fmt.Sprintf("%2d", tm.Hour()), true
-	case 'l':
-		hour := tm.Hour() % 12
-		if hour == 0 {
-			hour = 12
-		}
-		return fmt.Sprintf("%2d", hour), true
-	case 'u':
-		return strconv.Itoa(weekdayIndex(tm, true) + 1), true
-	case 'w':
-		return strconv.Itoa(int(tm.Weekday())), true
-	case 'G':
-		year, _ := tm.ISOWeek()
-		return strconv.Itoa(year), true
-	case 'V':
-		_, week := tm.ISOWeek()
-		return fmt.Sprintf("%02d", week), true
-	case 'U':
-		return fmt.Sprintf("%02d", (tm.YearDay()-1+7-int(tm.Weekday()))/7), true
-	case 'W':
-		return fmt.Sprintf("%02d", (tm.YearDay()-1+7-weekdayIndex(tm, true))/7), true
-	case 'n':
-		return "\n", true
-	case 't':
-		return "\t", true
-	default:
-		return "", false
-	}
-}
-
-// fnFormatDate implements dialects.GoogleSQL FORMAT_DATE/FORMAT_DATETIME/
-// FORMAT_TIMESTAMP(format, value). The format comes first, the reverse of dialects.MySQL
-// DATE_FORMAT.
-func fnFormatDate(args []driver.Value) (driver.Value, error) {
-	format, ok := toString(args[0])
-	if !ok {
-		return nil, nil
-	}
-	tm, ok := toStringTime(args[1])
-	if !ok {
-		return nil, nil
-	}
-	return strftimeRender(tm, format), nil
-}
-
-// fnParseDate implements dialects.GoogleSQL PARSE_DATE(format, text) and fnParseTimestamp
-// its datetime counterparts, returning NULL when the text does not match.
-func fnParseDate(args []driver.Value) (driver.Value, error) {
-	return parseWithStrftime(args, layoutDateOnly)
-}
-
-func fnParseTimestamp(args []driver.Value) (driver.Value, error) {
-	return parseWithStrftime(args, layoutDateTime)
-}
-
-func parseWithStrftime(args []driver.Value, out string) (driver.Value, error) {
-	format, ok1 := toString(args[0])
-	s, ok2 := toString(args[1])
-	if !ok1 || !ok2 {
-		return nil, nil
-	}
-	tm, ok := parseLayout(strftimeLayout(format), s)
-	if !ok {
-		return nil, nil
-	}
-	return tm.Format(out), nil
-}
-
-// unixScale builds UNIX_SECONDS/UNIX_MILLIS/UNIX_MICROS, which differ only in
-// how many sub-second units they report.
-func unixScale(perSecond int64) scalarFn {
-	return func(args []driver.Value) (driver.Value, error) {
-		tm, ok := toStringTime(args[0])
-		if !ok {
-			return nil, nil
-		}
-		return tm.Unix()*perSecond + int64(tm.Nanosecond())/(1000000000/perSecond), nil
-	}
-}
-
-// fromUnixScale builds TIMESTAMP_SECONDS/TIMESTAMP_MILLIS/TIMESTAMP_MICROS, the
-// inverses of unixScale.
-func fromUnixScale(perSecond int64) scalarFn {
-	return func(args []driver.Value) (driver.Value, error) {
-		n, ok := toInt(args[0])
-		if !ok {
-			return nil, nil
-		}
-		nanos := (n % perSecond) * (1000000000 / perSecond)
-		return time.Unix(n/perSecond, nanos).UTC().Format(layoutDateTime), nil
-	}
-}
-
-// fnToHex implements dialects.GoogleSQL TO_HEX(bytes): the lowercase hexadecimal form of
-// the value's bytes.
-func fnToHex(args []driver.Value) (driver.Value, error) {
-	switch x := args[0].(type) {
-	case nil:
-		return nil, nil
-	case []byte:
-		return hex.EncodeToString(x), nil
-	default:
-		s, ok := toString(args[0])
-		if !ok {
-			return nil, nil
-		}
-		return hex.EncodeToString([]byte(s)), nil
-	}
-}
-
-// fnPostgresToHex implements dialects.PostgreSQL to_hex(n): the lowercase hexadecimal
-// digits of an integer, with a negative read as a 64-bit two's complement value
-// the way dialects.PostgreSQL reads it. dialects.PostgreSQL has no string form of the function, so
-// a value that names no integer is refused rather than hexed as text, which is
-// what dialects.GoogleSQL's TO_HEX does with its bytes.
-func fnPostgresToHex(args []driver.Value) (driver.Value, error) {
-	if args[0] == nil {
-		return nil, nil
-	}
-	n, ok, err := postgresHexArgument(args[0])
-	if err != nil || !ok {
-		return nil, err
-	}
-	// The unsigned reading is the point: dialects.PostgreSQL answers the digits of the
-	// 64-bit two's complement value, so to_hex(-1) is sixteen f's.
-	return strconv.FormatUint(uint64(n), 16), nil //nolint:gosec // the two's complement reading is what dialects.PostgreSQL prints for a negative
-}
-
-// postgresHexArgument reads the integer to_hex converts. ok is false when the
-// value carries no value at all; err is set when it names something that is not
-// an integer, which dialects.PostgreSQL has no to_hex for.
-func postgresHexArgument(v driver.Value) (int64, bool, error) {
-	switch x := v.(type) {
-	case int64:
-		return x, true, nil
-	case float64:
-		if math.IsNaN(x) || math.IsInf(x, 0) || x != math.Trunc(x) {
-			return 0, false, fmt.Errorf("%w: to_hex expects an integer, got %v", ErrInvalidCast, x)
-		}
-		return int64(x), true, nil
-	}
-	s, ok := toString(v)
-	if !ok {
-		return 0, false, nil
-	}
-	n, err := strconv.ParseInt(strings.TrimSpace(s), 10, 64)
-	if err != nil {
-		return 0, false, fmt.Errorf("%w: to_hex expects an integer, got %q", ErrInvalidCast, s)
-	}
-	return n, true, nil
-}
-
-// fnMySQLFormat implements dialects.MySQL FORMAT(x, d): x rounded to d decimal places and
-// written with a comma every three digits. dialects.SQLite has a format() of its own, an
-// alias of printf, so an untranslated call answered the first argument expanded
-// as a format string instead.
-func fnMySQLFormat(args []driver.Value) (driver.Value, error) {
-	d, ok2 := toInt(args[1])
-	if !ok2 {
-		return nil, nil
-	}
-	// dialects.MySQL reads a negative number of decimal places as none, and caps the
-	// count at the 30 its DECIMAL type holds.
-	if d < 0 {
-		d = 0
-	}
-	if d > mysqlFormatMaxDecimals {
-		d = mysqlFormatMaxDecimals
-	}
-	// An integer is grouped from its own digits. Widening it to a float64 first
-	// lost the last digits of anything past 2^53, so FORMAT on a large id column
-	// printed a number one or two away from the one stored.
-	if n, isInt := args[0].(int64); isInt {
-		return formatGrouped(strconv.FormatInt(n, 10), int(d)), nil
-	}
-	x, ok1 := mysqlNumericArgument(args[0])
-	if !ok1 {
-		return nil, nil
-	}
-	if math.IsNaN(x) || math.IsInf(x, 0) {
-		return nil, nil
-	}
-	text := mysqlFormatDecimal(x, int(d))
-	sign := ""
-	if strings.HasPrefix(text, "-") {
-		sign, text = "-", text[1:]
-	}
-	whole, fraction, hasFraction := strings.Cut(text, ".")
-	out := sign + groupThousands(whole)
-	if hasFraction {
-		out += "." + fraction
-	}
-	return out, nil
-}
-
-// formatGrouped writes an integer already spelled in decimal the way dialects.MySQL's
-// FORMAT writes it: grouped in threes, with decimals decimal places of zeros
-// after it.
-func formatGrouped(digits string, decimals int) string {
-	sign := ""
-	if strings.HasPrefix(digits, "-") {
-		sign, digits = "-", digits[1:]
-	}
-	out := sign + groupThousands(digits)
-	if decimals > 0 {
-		out += "." + strings.Repeat("0", decimals)
-	}
-	return out
-}
-
-// mysqlFormatMaxDecimals is the number of decimal places dialects.MySQL FORMAT keeps at
-// most, which is the scale of its DECIMAL type.
-// mysqlFormatDecimal is the number MySQL's FORMAT groups: the shortest decimal
-// that reads back as the same double, rounded to decimals places with a half
-// going to the even neighbor.
-//
-// Both halves of that are the engine's. Formatting the double itself printed
-// the exact value of the binary number, so FORMAT(1e100, 2) came out as
-// 10000000000000000159028911097599180468360808563945281389781327557747838772170381060813469985856815104.00
-// where MySQL prints 1 followed by 100 zeros: the shortest decimal is the one
-// the caller wrote, so what fills the places past it is zeros rather than the
-// double's own digits. And a half goes to the even neighbor for a
-// floating-point argument, the rule ROUND is on in MySQL and PostgreSQL, so
-// FORMAT(2.5e0, 0) is 2 and FORMAT(3.5e0, 0) is 4.
-//
-// The rounding runs on the digits rather than on the value, because scaling a
-// double by a power of ten is what loses the tie: 2.675 is a shade below itself
-// in binary, so 2.675 * 100 is 267.49999999999997 and rounds down, where MySQL
-// answers 2.68.
-func mysqlFormatDecimal(x float64, decimals int) string {
-	text := strconv.FormatFloat(x, 'f', -1, 64)
-	sign := ""
-	if strings.HasPrefix(text, "-") {
-		sign, text = "-", text[1:]
-	}
-	whole, fraction, _ := strings.Cut(text, ".")
-	if len(fraction) < decimals {
-		fraction += strings.Repeat("0", decimals-len(fraction))
-	}
-	digits := whole + fraction[:decimals]
-	if stepsUpHalfToEven(digits, fraction[decimals:]) {
-		digits = incrementDecimal(digits)
-	}
-	if len(digits) <= decimals {
-		digits = strings.Repeat("0", decimals+1-len(digits)) + digits
-	}
-	out := sign + digits[:len(digits)-decimals]
-	if decimals > 0 {
-		out += "." + digits[len(digits)-decimals:]
-	}
-	return out
-}
-
-// stepsUpHalfToEven reports whether the digits being cut off ask the kept ones
-// to be stepped up: more than a half does, and exactly a half does only when
-// the last kept digit is odd.
-func stepsUpHalfToEven(kept, cut string) bool {
-	if cut == "" || cut[0] < '5' {
-		return false
-	}
-	if cut[0] > '5' {
-		return true
-	}
-	for i := 1; i < len(cut); i++ {
-		if cut[i] != '0' {
-			return true
-		}
-	}
-	last := byte('0')
-	if kept != "" {
-		last = kept[len(kept)-1]
-	}
-	return (last-'0')%2 == 1
-}
-
-// incrementDecimal adds one to a run of decimal digits, growing it by a place
-// when every digit was a nine.
-func incrementDecimal(digits string) string {
-	out := []byte(digits)
-	for i := len(out) - 1; i >= 0; i-- {
-		if out[i] != '9' {
-			out[i]++
-			return string(out)
-		}
-		out[i] = '0'
-	}
-	return "1" + string(out)
-}
-
-const mysqlFormatMaxDecimals = 30
-
-// fnGoogleSQLFormat implements dialects.GoogleSQL FORMAT(format, ...). The verbs it
-// shares with printf are handed to Sprintf, so they answer what dialects.SQLite's printf
-// answered before; %t and %T are BigQuery's own and are printed here. Left to
-// dialects.SQLite they made the whole call NULL, since printf answers NULL for a format
-// string holding a verb it does not know.
-func fnGoogleSQLFormat(args []driver.Value) (driver.Value, error) {
-	if len(args) == 0 {
-		return nil, errors.New("dialect: FORMAT expects a format string")
-	}
-	format, ok := toString(args[0])
-	if !ok {
-		return nil, nil
-	}
-	rest, next := args[1:], 0
-	take := func() driver.Value {
-		if next >= len(rest) {
-			return nil
-		}
-		v := rest[next]
-		next++
-		return v
-	}
-	// The result is built in a byte slice rather than a strings.Builder so the
-	// printf verbs can be appended in place: Appendf writes into it, where
-	// Sprintf would allocate a string per verb only to copy it in.
-	out := make([]byte, 0, len(format)+16)
-	for i := 0; i < len(format); {
-		if format[i] != '%' {
-			out = append(out, format[i])
-			i++
-			continue
-		}
-		spec, verb, end := scanFormatSpec(format, i)
-		switch {
-		case end < 0:
-			// A trailing "%" with no verb after it stands for itself.
-			out = append(out, format[i:]...)
-			i = len(format)
-		case verb == '%':
-			out = append(out, '%')
-			i = end
-		case verb == 't' || verb == 'T':
-			out = append(out, googlesqlPrintValue(take(), verb == 'T')...)
-			i = end
-		default:
-			operands := make([]any, 0, strings.Count(spec, "*")+1)
-			for range strings.Count(spec, "*") {
-				operands = append(operands, formatOperand('d', take()))
-			}
-			operands = append(operands, formatOperand(verb, take()))
-			out = fmt.Appendf(out, goFormatSpec(spec, verb), operands...)
-			i = end
-		}
-	}
-	return string(out), nil
-}
-
-// scanFormatSpec reads the conversion specification that starts at the "%" at
-// index start, returning the specification text, its verb, and the index just
-// past it. end is -1 when the string ends before a verb.
-func scanFormatSpec(format string, start int) (string, byte, int) {
-	i := start + 1
-	for i < len(format) && strings.IndexByte("+-# 0123456789.*'", format[i]) >= 0 {
-		i++
-	}
-	if i >= len(format) {
-		return format[start:], 0, -1
-	}
-	return format[start : i+1], format[i], i + 1
-}
-
-// goFormatSpec adapts a printf specification to the one Go understands: the
-// apostrophe flag, which asks C for digit grouping, has no Go equivalent and is
-// dropped, and %i is the spelling of %d that C printf accepts.
-func goFormatSpec(spec string, verb byte) string {
-	spec = strings.ReplaceAll(spec, "'", "")
-	if verb == 'i' {
-		spec = spec[:len(spec)-1] + "d"
-	}
-	return spec
-}
-
-// formatOperand coerces a value to the Go type the verb prints, the way dialects.SQLite's
-// printf coerces its own arguments: a missing or unreadable value prints as the
-// zero of the verb's type rather than as a Go error string.
-func formatOperand(verb byte, v driver.Value) any {
-	switch verb {
-	case 'd', 'i', 'o', 'b', 'x', 'X', 'c', 'U':
-		n, _ := toInt(v)
-		return n
-	case 'e', 'E', 'f', 'F', 'g', 'G':
-		f, _ := toFloat(v)
-		return f
-	default:
-		s, _ := toString(v)
-		return s
-	}
-}
-
-// nullText is how a value that has none is spelled where SQL spells it out:
-// dialects.GoogleSQL's FORMAT prints it, dialects.MySQL's QUOTE answers it, dialects.PostgreSQL's
-// quote_nullable answers it, and the keyword tables in this package hold it.
-const nullText = "NULL"
-
-// googlesqlPrintValue implements dialects.GoogleSQL's %t and %T: the printable form of a
-// value, and the literal that would produce it. A boolean reaches here as the
-// integer dialects.SQLite stores, so it prints as 0 or 1 rather than as false or true.
-func googlesqlPrintValue(v driver.Value, literal bool) string {
-	switch x := v.(type) {
-	case nil:
-		return nullText
-	case int64:
-		return strconv.FormatInt(x, 10)
-	case float64:
-		return strconv.FormatFloat(x, 'g', -1, 64)
-	case bool:
-		return strconv.FormatBool(x)
-	}
-	s, ok := toString(v)
-	if !ok {
-		return nullText
-	}
-	if literal {
-		return strconv.Quote(s)
-	}
-	return s
-}
-
-// fnIsNaN implements dialects.GoogleSQL IS_NAN(x).
-func fnIsNaN(args []driver.Value) (driver.Value, error) {
-	f, ok := toFloat(args[0])
-	if !ok {
-		// A NULL argument is NULL; a non-numeric one is simply not NaN.
-		if _, present := toString(args[0]); !present {
-			return nil, nil
-		}
-		return int64(0), nil
-	}
-	return boolToInt(math.IsNaN(f)), nil
-}
-
-// safeArith builds the dialects.GoogleSQL SAFE_ADD/SAFE_SUBTRACT/SAFE_MULTIPLY family:
-// the operation, or NULL where the source dialect would raise an overflow. Two
-// integer arguments stay in int64 so overflow is detectable; anything else falls
-// back to float64, where the result is already an infinity rather than an error.
-func safeArith(intOp func(a, b int64) (int64, bool), floatOp func(a, b float64) float64) scalarFn {
-	return func(args []driver.Value) (driver.Value, error) {
-		if ai, ok := args[0].(int64); ok {
-			if bi, ok := args[1].(int64); ok {
-				res, fine := intOp(ai, bi)
-				if !fine {
-					return nil, nil
-				}
-				return res, nil
-			}
-		}
-		a, ok1 := toFloat(args[0])
-		b, ok2 := toFloat(args[1])
-		if !ok1 || !ok2 {
-			return nil, nil
-		}
-		res := floatOp(a, b)
-		if math.IsInf(res, 0) || math.IsNaN(res) {
-			return nil, nil
-		}
-		return res, nil
-	}
-}
-
-func safeAddInt(a, b int64) (int64, bool) {
-	sum := a + b
-	// The sum overflowed when both operands share a sign that the result lost.
-	if (a > 0 && b > 0 && sum < 0) || (a < 0 && b < 0 && sum >= 0) {
-		return 0, false
-	}
-	return sum, true
-}
-
-func safeSubInt(a, b int64) (int64, bool) {
-	if b == math.MinInt64 {
-		// Negating the minimum has no int64 representation, so handle it as an
-		// addition that cannot be expressed either.
-		if a >= 0 {
-			return 0, false
-		}
-		return a - b, true
-	}
-	return safeAddInt(a, -b)
-}
-
-func safeMulInt(a, b int64) (int64, bool) {
-	if a == 0 || b == 0 {
-		return 0, true
-	}
-	product := a * b
-	if product/b != a || (a == -1 && b == math.MinInt64) || (b == -1 && a == math.MinInt64) {
-		return 0, false
-	}
-	return product, true
-}
-
-// fnSafeNegate implements dialects.GoogleSQL SAFE_NEGATE(x).
-func fnSafeNegate(args []driver.Value) (driver.Value, error) {
-	if n, ok := args[0].(int64); ok {
-		if n == math.MinInt64 {
-			return nil, nil
-		}
-		return -n, nil
-	}
-	f, ok := toFloat(args[0])
-	if !ok {
-		return nil, nil
-	}
-	return -f, nil
-}
-
-// fnGenerateUUID implements dialects.GoogleSQL GENERATE_UUID: a random RFC 4122 v4 UUID.
-func fnGenerateUUID(_ []driver.Value) (driver.Value, error) {
-	var b [16]byte
-	if _, err := rand.Read(b[:]); err != nil {
-		return nil, fmt.Errorf("dialect: generate_uuid: %w", err)
-	}
-	b[6] = (b[6] & 0x0f) | 0x40 // version 4
-	b[8] = (b[8] & 0x3f) | 0x80 // variant 10
-	return fmt.Sprintf("%x-%x-%x-%x-%x", b[0:4], b[4:6], b[6:8], b[8:10], b[10:16]), nil
-}
-
-// --- time parsing ---
-
-// timeLayouts are tried in order by parseTime, covering the common SQL date and
-// datetime shapes. The Z07:00 pair reads a timezone suffix — a trailing Z or an
-// offset like +09:00 — which BigQuery's TIMESTAMP literals carry.
-var timeLayouts = []string{
-	"2006-01-02 15:04:05.999999999",
-	"2006-01-02 15:04:05",
-	"2006-01-02 15:04:05.999999999Z07:00",
-	"2006-01-02T15:04:05",
-	"2006-01-02T15:04:05.999999999Z07:00",
-	"2006-01-02 15:04",
-	"2006-01-02",
-	"2006/01/02 15:04:05",
-	"2006/01/02",
-	"15:04:05",
-	"15:04:05.999999999",
-}
-
-// parseTime parses a date/time string against the supported layouts. A value
-// with a timezone suffix is normalized to UTC, which is the timezone BigQuery
-// extracts fields in by default.
-func parseTime(s string) (time.Time, bool) {
-	s = strings.TrimSpace(s)
-	for _, layout := range timeLayouts {
-		if tm, err := time.Parse(layout, s); err == nil {
-			return tm.UTC(), true
-		}
-	}
-	return time.Time{}, false
-}
-
-// formatFloatText renders a REAL as text the way the engines do. Go's shortest
-// 'g' switches to exponent notation once the decimal exponent reaches 6, so a
-// plain 1234567.5 came out of every function that reads its argument as text as
-// "1.2345675e+06", where dialects.SQLite's own concatenation, dialects.MySQL and dialects.PostgreSQL all
-// write 1234567.5. Between 1e-4 and 1e15 the three agree on plain notation and
-// the value is written plainly; outside that band they disagree with each other
-// and the shortest form is kept.
-func formatFloatText(f float64) string {
-	if abs := math.Abs(f); abs != 0 && (abs < 1e-4 || abs >= 1e15) {
-		return strconv.FormatFloat(f, 'g', -1, 64)
-	}
-	return strconv.FormatFloat(f, 'f', -1, 64)
-}
-
-// formatFloatTextMySQL renders a REAL the way MySQL writes one in a string
-// context, which is not how formatFloatText writes it: MySQL has no plus sign
-// and no padding in an exponent, and it keeps the plain spelling out to 1e14
-// and down to 1e-15, so 1e308 is "1e308" rather than "1e+308" and 1e-5 is
-// "0.00001" rather than "1e-05".
-//
-// Which of the two a value takes turns on where the decimal point sits relative
-// to the digits rather than on the length alone, which is why 1234567890123456
-// is written with an exponent and 1234567890123456.8, one digit longer, is
-// written plainly. TestMySQLWritesARealTheWayTheEngineDoes holds the values
-// this was read from.
-func formatFloatTextMySQL(f float64) string {
-	if math.IsInf(f, 0) || math.IsNaN(f) {
-		// Neither is a value SQLite can hold in a column loaded from a file,
-		// and MySQL has no spelling for them; the shared one applies.
-		return formatFloatText(f)
-	}
-	digits, point := shortestDecimal(f)
-	if point <= -15 || (point >= 16 && point >= len(digits)) || plainTextWidth(len(digits), point) > 24 {
-		return exponentTextMySQL(f, digits, point)
-	}
-	return strconv.FormatFloat(f, 'f', -1, 64)
-}
-
-// shortestDecimal is the shortest run of digits that reads back as f, and where
-// the decimal point sits in it: f is 0.<digits> times ten to the point.
-func shortestDecimal(f float64) (digits string, point int) {
-	written := strconv.FormatFloat(math.Abs(f), 'e', -1, 64)
-	mantissa, exponent, _ := strings.Cut(written, "e")
-	e, err := strconv.Atoi(exponent)
-	if err != nil {
-		// FormatFloat with 'e' always writes an exponent, so this cannot
-		// happen; answering the mantissa alone keeps a caller from a panic.
-		return strings.Replace(mantissa, ".", "", 1), 1
-	}
-	return strings.Replace(mantissa, ".", "", 1), e + 1
-}
-
-// plainTextWidth is how many characters the plain spelling of a value takes,
-// counting the "0." a value below one opens with and the point a fraction needs.
-func plainTextWidth(length, point int) int {
-	if point <= 0 {
-		return length - point + 2
-	}
-	if point < length {
-		return length + 1
-	}
-	return point
-}
-
-// exponentTextMySQL writes a value in MySQL's exponent spelling: one digit, the
-// rest after a point, then "e" and the exponent with a sign only when it is
-// negative.
-func exponentTextMySQL(f float64, digits string, point int) string {
-	var b strings.Builder
-	if math.Signbit(f) {
-		b.WriteByte('-')
-	}
-	b.WriteByte(digits[0])
-	if len(digits) > 1 {
-		b.WriteByte('.')
-		b.WriteString(digits[1:])
-	}
-	b.WriteByte('e')
-	b.WriteString(strconv.Itoa(point - 1))
-	return b.String()
-}
-
-// fnMySQLText writes a value the way MySQL writes one a string function reads.
-// Only a REAL is rewritten, so a blob stays a blob and an integer stays an
-// integer for the call it is handed to.
-func fnMySQLText(args []driver.Value) (driver.Value, error) {
-	if f, ok := args[0].(float64); ok {
-		return formatFloatTextMySQL(f), nil
-	}
-	return args[0], nil
-}
-
-// mysqlTextAll is mysqlTextArgs over every argument.
-func mysqlTextAll(fn scalarFn) scalarFn { return mysqlTextArgs(fn) }
-
-// mysqlTextFrom is mysqlTextArgs over every argument from first onward, for the
-// helpers whose text arguments are a tail of any length.
-func mysqlTextFrom(fn scalarFn, first int) scalarFn {
-	return func(args []driver.Value) (driver.Value, error) {
-		for i := first; i < len(args); i++ {
-			if f, ok := args[i].(float64); ok {
-				args[i] = formatFloatTextMySQL(f)
-			}
-		}
-		return fn(args)
-	}
-}
-
-// mysqlTextArgs returns fn with the arguments at the given positions -- every
-// argument when none are named -- written the way MySQL writes a value a string
-// function reads. Only a REAL is rewritten: MySQL spells an integer, a string
-// and a blob the way this package already does.
-//
-// The helper it wraps stays shared rather than gaining a MySQL twin, because
-// PostgreSQL and GoogleSQL refuse a float to these functions outright: there is
-// no second answer to keep.
-func mysqlTextArgs(fn scalarFn, positions ...int) scalarFn {
-	return func(args []driver.Value) (driver.Value, error) {
-		if len(positions) == 0 {
-			for i, arg := range args {
-				if f, ok := arg.(float64); ok {
-					args[i] = formatFloatTextMySQL(f)
-				}
-			}
-			return fn(args)
-		}
-		for _, i := range positions {
-			if i < 0 || i >= len(args) {
-				continue
-			}
-			if f, ok := args[i].(float64); ok {
-				args[i] = formatFloatTextMySQL(f)
-			}
-		}
-		return fn(args)
-	}
-}
-
-// toStringTime coerces a value to a time, accepting both strings and time.Time.
-func toStringTime(v driver.Value) (time.Time, bool) {
-	if tm, ok := v.(time.Time); ok {
-		return tm, true
-	}
-	s, ok := toString(v)
-	if !ok {
-		return time.Time{}, false
-	}
-	return parseTime(s)
 }
